Improve status message when pebble is not ready (#325)
* Improve status message when pebble is not ready

This commit changes the status message when pebble is not
read to "Configuring Prometheus" instead of "Waiting for pebble
ready" because the latter is not meaningful to the system
adminstrator. Also the status has been changed from WaitingStatus to
MaintenanceStatus.

Closes #267

* Integration tests are now specific when adding relations

Recent commits added self monitoring relations to charms in the
observability charms themselves. As a result of this observability
charms that can now have more than one relation type with the Prometheus
charm. Hence adding or removing relations in integration tests needs to
be very specific about which relation type is being added or removed.
Without this the integration tests fail due to ambiguous relation.
This commit fixes these issues.

Closes #326

* Fix scaling parameter argument

This commit fixes a minor bug. The Python libjuju Scaling API used
set scale change to zero rather than scale to zero.